test: add conformance harness + fix protocol validation for h3-test compliance
- Add ConformanceHarness supporting full agent loop (tool_call→result→text→end)
- Add examples/conformance HTTP server for running h3-test against
- Add testbed/fixtures/ with sample JSON request/response payloads
- Relax ProcessRequest validation (timestamp/user/config optional per h3-test)
- Fix Result.DurationMs int→float64 to match h3-test wire format
- Fix cancel endpoint 204→200 to match h3-test expectations
- h3-test score: 42/43 PASS (≥ 40/43 target)

+++ protocol/types.go
@@ -210,7 +210,7 @@ type Result struct {
 	Type        ResultType `json:"type"`
 	ToolName    string     `json:"tool_name,omitempty"`
 	Data        any        `json:"data,omitempty"`
-	DurationMs  int        `json:"duration_ms,omitempty"`
+	DurationMs  float64    `json:"duration_ms,omitempty"`
 	Success     bool       `json:"success"`
 }
